+++ include/TGUI/Widgets/CanvasBase.hpp
@@ -57,14 +57,14 @@ TGUI_MODULE_EXPORT namespace tgui
         ///
         /// By default, mouse events are NOT ignored.
         /////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////
-        void ignoreMouseEvents(bool ignore = true);
+        TGUI_DEPRECATED("Use setIgnoreMouseEvents instead") void ignoreMouseEvents(bool ignore = true);
 
         /////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////
         /// @brief Returns whether the widget is ignoring mouse events and letting them pass to the widgets behind it
         ///
         /// @return Are mouse events ignored by this widget?
         /////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////
-        TGUI_NODISCARD bool isIgnoringMouseEvents() const;
+        TGUI_DEPRECATED("Use getIgnoreMouseEvents instead") TGUI_NODISCARD bool isIgnoringMouseEvents() const;
 
         /////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////
         /// @brief Returns whether the widget can gain focus
@@ -96,7 +96,7 @@ TGUI_MODULE_EXPORT namespace tgui
         /////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////
     private:
 
-        bool m_ignoringMouseEvents = false;
+        bool m_ignoringMouseEvents = false; // TGUI_NEXT: Remove this property
     };
 }