+++ src/Widgets/CanvasBase.cpp
@@ -85,7 +85,7 @@ namespace tgui
         ClickableWidget::load(node, renderers);
 
         if (node->propertyValuePairs[U"IgnoreMouseEvents"])
-            ignoreMouseEvents(Deserializer::deserialize(ObjectConverter::Type::Bool, node->propertyValuePairs[U"IgnoreMouseEvents"]->value).getBool());
+            m_ignoringMouseEvents = Deserializer::deserialize(ObjectConverter::Type::Bool, node->propertyValuePairs[U"IgnoreMouseEvents"]->value).getBool();
     }
 
     /////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////